Resolve core guidelines warnings in x64 builds

+++ samples/convert.c/main.c
@@ -99,7 +99,7 @@ static void *bmp_read_pixel_data(FILE *fp, uint32_t offset, const bmp_dib_header
     if (fseek(fp, offset, SEEK_SET))
         return NULL;
 
-    *buffer_size = header->height * header->width * 3;
+    *buffer_size = (size_t)header->height * header->width * 3;
     void *buffer = malloc(*buffer_size);
     if (buffer)
     {
